+++ internal/graphapi/testharness/harness.go
@@ -3,32 +3,21 @@
 package testharness
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"fmt"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"path/filepath"
 	"runtime"
-	"sync"
 	"testing"
 	"time"
 
 	"github.com/99designs/gqlgen/graphql"
 	"github.com/99designs/gqlgen/graphql/handler"
 	"github.com/99designs/gqlgen/graphql/handler/transport"
-	"github.com/gqlgo/gqlgenc/clientv2"
 	"github.com/mcuadros/go-defaults"
-	"github.com/pdfcpu/pdfcpu/pkg/api"
 	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"github.com/stripe/stripe-go/v86"
-	"github.com/vektah/gqlparser/v2/gqlerror"
-	"gotest.tools/v3/assert"
 
 	echo "github.com/theopenlane/echox"
 	"github.com/theopenlane/iam/fgax"
@@ -41,7 +30,6 @@ import (
 	"github.com/theopenlane/utils/testutils"
 	"github.com/theopenlane/utils/ulids"
 
-	"github.com/theopenlane/core/common/enums"
 	"github.com/theopenlane/core/common/storagetypes"
 	"github.com/theopenlane/core/v2/fga/fgaversion"
 	"github.com/theopenlane/core/v2/internal/ent/entconfig"
@@ -65,16 +53,13 @@ import (
 	coreutils "github.com/theopenlane/core/v2/internal/testutils"
 	testint "github.com/theopenlane/core/v2/internal/testutils/integrations"
 	"github.com/theopenlane/core/v2/internal/workflows/engine"
-	"github.com/theopenlane/core/v2/pkg/entitlements"
 	"github.com/theopenlane/core/v2/pkg/entitlements/mocks"
 	"github.com/theopenlane/core/v2/pkg/gala"
 	"github.com/theopenlane/core/v2/pkg/logx"
 	authmw "github.com/theopenlane/core/v2/pkg/middleware/auth"
-	pkgobjects "github.com/theopenlane/core/v2/pkg/objects"
 	mock_shared "github.com/theopenlane/core/v2/pkg/objects/mocks"
 	"github.com/theopenlane/core/v2/pkg/objects/storage"
 	"github.com/theopenlane/core/v2/pkg/summarizer"
-	mockprovider "github.com/theopenlane/newman/providers/mock"
 
 	// import generated runtime which is required to prevent cyclical dependencies
 	"github.com/theopenlane/core/v2/internal/ent/generated/privacy"
@@ -101,17 +86,14 @@ const (
 
 // GraphTestSuite handles the setup and teardown between tests
 type GraphTestSuite struct {
-	Client               *Client
-	TF                   *testutils.TestFixture
-	OFGATF               *fgatest.OpenFGATestFixture
-	StripeMockBackend    *mocks.MockStripeBackend
-	CacheRefreshServer   *httptest.Server
-	GalaRuntime          *gala.Gala
-	IntegrationsRT       *intruntime.Runtime
-	WorkflowEngine       *engine.WorkflowEngine
-	WorkflowListenersMu  sync.Mutex
-	WorkflowListenerRefs int
-	WorkflowListenerIDs  []gala.ListenerID
+	Client             *Client
+	TF                 *testutils.TestFixture
+	OFGATF             *fgatest.OpenFGATestFixture
+	StripeMockBackend  *mocks.MockStripeBackend
+	CacheRefreshServer *httptest.Server
+	GalaRuntime        *gala.Gala
+	IntegrationsRT     *intruntime.Runtime
+	WorkflowEngine     *engine.WorkflowEngine
 }
 
 // Client contains all the clients the test need to interact with
@@ -125,36 +107,6 @@ type Client struct {
 	ObjectStore        *objects.Service
 	MockProvider       *mock_shared.MockProvider
 	DeletedStorageKeys *DeletedKeys
-}
-
-// DeletedKeys is a concurrency safe set of the storage keys removed from object storage, the gala
-// workers deleting them run on their own goroutines while the test asserts on the main one
-type DeletedKeys struct {
-	mu   sync.Mutex
-	keys map[string]struct{}
-}
-
-// NewDeletedKeys returns an empty set
-func NewDeletedKeys() *DeletedKeys {
-	return &DeletedKeys{keys: map[string]struct{}{}}
-}
-
-// Add records a storage key that was deleted
-func (d *DeletedKeys) Add(key string) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	d.keys[key] = struct{}{}
-}
-
-// Has reports whether the given storage key was deleted
-func (d *DeletedKeys) Has(key string) bool {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	_, ok := d.keys[key]
-
-	return ok
 }
 
 // repoRoot resolves the repository root from this file's own location so paths below do
@@ -438,526 +390,6 @@ func (suite *GraphTestSuite) TearDownSuite(t *testing.T) {
 	if suite.CacheRefreshServer != nil {
 		suite.CacheRefreshServer.Close()
 	}
-}
-
-// AcquireWorkflowRuntime shares one workflow listener registration across parallel tests
-func (suite *GraphTestSuite) AcquireWorkflowRuntime(t *testing.T) (*engine.WorkflowEngine, *gala.Gala) {
-	t.Helper()
-
-	suite.WorkflowListenersMu.Lock()
-	if suite.WorkflowListenerRefs == 0 {
-		listenerIDs, err := gala.Register(suite.GalaRuntime, hooks.WorkflowListeners()...)
-		if err != nil {
-			suite.WorkflowListenersMu.Unlock()
-			RequireNoError(t, err)
-
-			return nil, nil
-		}
-
-		suite.WorkflowListenerIDs = listenerIDs
-	}
-
-	suite.WorkflowListenerRefs++
-	workflowEngine := suite.WorkflowEngine
-	runtime := suite.GalaRuntime
-	suite.WorkflowListenersMu.Unlock()
-
-	t.Cleanup(func() {
-		suite.ReleaseWorkflowRuntime(t)
-	})
-
-	return workflowEngine, runtime
-}
-
-// ReleaseWorkflowRuntime removes workflow listeners after the last borrowing test
-func (suite *GraphTestSuite) ReleaseWorkflowRuntime(t *testing.T) {
-	t.Helper()
-
-	suite.WorkflowListenersMu.Lock()
-	defer suite.WorkflowListenersMu.Unlock()
-
-	if suite.WorkflowListenerRefs == 0 {
-		return
-	}
-
-	suite.WorkflowListenerRefs--
-	if suite.WorkflowListenerRefs > 0 {
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), gala.DefaultSoftStopTimeout)
-	defer cancel()
-
-	err := suite.GalaRuntime.RemoveListeners(ctx, suite.WorkflowListenerIDs...)
-	RequireNoError(t, err)
-	suite.WorkflowListenerIDs = nil
-}
-
-// WaitForEvents blocks until runnable and in-flight Gala jobs complete
-func (suite *GraphTestSuite) WaitForEvents() {
-	if err := suite.GalaRuntime.WaitIdle(context.Background()); err != nil {
-		panic(err)
-	}
-}
-
-func WaitForGala(t *testing.T, runtime *gala.Gala) {
-	t.Helper()
-	assert.NilError(t, runtime.WaitIdle(t.Context()))
-}
-
-// MockEmailSender extracts the mock email sender from the integration runtime
-func (suite *GraphTestSuite) MockEmailSender() *mockprovider.EmailSender {
-	rc, ok := suite.IntegrationsRT.Registry().RuntimeClient(emaildef.DefinitionID.ID())
-	if !ok {
-		panic("email runtime client not found")
-	}
-
-	ms := emaildef.MockSenderFromClient(rc)
-	if ms == nil {
-		panic("mock sender not found")
-	}
-
-	return ms
-}
-
-func (suite *GraphTestSuite) EnableGalaForTestSuite(t *testing.T) {
-	t.Helper()
-
-	if suite.GalaRuntime != nil {
-		return
-	}
-
-	runtime, err := gala.NewGala(context.Background(), gala.Config{
-		ConnectionURI:     suite.TF.URI,
-		QueueName:         fmt.Sprintf("graphapi_test_%d", time.Now().UnixNano()),
-		WorkerCount:       1,
-		RunMigrations:     true,
-		FetchCooldown:     time.Millisecond,
-		FetchPollInterval: 10 * time.Millisecond,
-	})
-	require.NoError(t, err)
-
-	require.NoError(t, runtime.Attach(
-		gala.WithValue(runtime),
-		gala.WithValue(suite.Client.DB),
-		gala.WithValue(suite.Client.DB.EntitlementManager),
-	))
-
-	_, err = gala.Register(runtime, hooks.EntitlementListeners()...)
-	require.NoError(t, err)
-
-	err = runtime.StartWorkers(context.Background())
-	require.NoError(t, err)
-
-	suite.GalaRuntime = runtime
-
-	t.Cleanup(func() {
-		if suite.GalaRuntime == nil {
-			return
-		}
-
-		err := suite.GalaRuntime.StopWorkers(context.Background())
-		require.NoError(t, err)
-
-		err = suite.GalaRuntime.Close()
-		require.NoError(t, err)
-
-		suite.GalaRuntime = nil
-	})
-}
-
-// ExpectUpload sets up the mock object store to expect an upload and related operations
-func ExpectUpload(t *testing.T, mockProvider *mock_shared.MockProvider, expectedUploads []graphql.Upload) {
-	assert.Assert(t, mockProvider != nil)
-
-	mockScheme := "file://"
-
-	for _, upload := range expectedUploads {
-		mockProvider.On("GetScheme").Return(&mockScheme).Once()
-		mockProvider.On("ProviderType").Return(storage.DiskProvider).Maybe()
-		mockProvider.On("Upload", mock.Anything, mock.Anything, mock.Anything).Return(&storage.UploadedMetadata{
-			FileMetadata: pkgobjects.FileMetadata{
-				Key:          "test-key",
-				Size:         upload.Size,
-				Folder:       "test-folder",
-				Bucket:       "test-bucket",
-				ContentType:  upload.ContentType,
-				ProviderType: storage.DiskProvider,
-				FullURI:      "file:///tmp/test-file",
-			},
-		}, nil).Once()
-
-		// Allow document hooks to download the just-uploaded content for parsing
-		mockProvider.On("Download", mock.Anything, mock.Anything, mock.Anything).Return(&storage.DownloadedMetadata{
-			File: TestPDFBytes(),
-			Size: upload.Size,
-		}, nil).Maybe()
-	}
-}
-
-// ExpectAttestedUpload sets up mock expectations for the attested PDF upload triggered by attestNDADocument
-func ExpectAttestedUpload(t *testing.T, mockProvider *mock_shared.MockProvider) {
-	assert.Assert(t, mockProvider != nil)
-
-	mockScheme := "file://"
-
-	mockProvider.On("GetScheme").Return(&mockScheme).Once()
-	mockProvider.On("ProviderType").Return(storage.DiskProvider).Maybe()
-	mockProvider.On("Upload", mock.Anything, mock.Anything, mock.Anything).Return(&storage.UploadedMetadata{
-		FileMetadata: pkgobjects.FileMetadata{
-			Key:          "test-key-attested",
-			Folder:       "test-folder",
-			Bucket:       "test-bucket",
-			ContentType:  "application/pdf",
-			ProviderType: storage.DiskProvider,
-			FullURI:      "file:///tmp/test-file-attested",
-		},
-	}, nil).Once()
-}
-
-func ExpectUploadWithTemplateKind(t *testing.T, mockProvider *mock_shared.MockProvider, expectedUploads []graphql.Upload, kind enums.TemplateKind) {
-	assert.Assert(t, mockProvider != nil)
-
-	mockScheme := "file://"
-
-	for _, upload := range expectedUploads {
-		mockProvider.On("GetScheme").Return(&mockScheme).Once()
-		mockProvider.On("ProviderType").Return(storage.DiskProvider).Maybe()
-		uploadOpts := mock.MatchedBy(func(opts *storage.UploadOptions) bool {
-			if opts == nil || opts.ProviderHints == nil || opts.ProviderHints.Metadata == nil {
-				return false
-			}
-
-			return opts.ProviderHints.Metadata[objects.TemplateKindMetadataKey] == kind.String()
-		})
-		mockProvider.On("Upload", mock.Anything, mock.Anything, uploadOpts).Return(&storage.UploadedMetadata{
-			FileMetadata: pkgobjects.FileMetadata{
-				Key:          "test-key",
-				Size:         upload.Size,
-				Folder:       "test-folder",
-				Bucket:       "test-bucket",
-				ContentType:  upload.ContentType,
-				ProviderType: storage.DiskProvider,
-				FullURI:      "file:///tmp/test-file",
-			},
-		}, nil).Once()
-
-		// Allow document hooks to download the just-uploaded content for parsing
-		mockProvider.On("Download", mock.Anything, mock.Anything, mock.Anything).Return(&storage.DownloadedMetadata{
-			File: TestPDFBytes(),
-			Size: upload.Size,
-		}, nil).Maybe()
-	}
-}
-
-// ExpectDelete sets up the mock object store to expect a delete and related operations
-func ExpectDelete(t *testing.T, mockProvider *mock_shared.MockProvider, expectedUploads []graphql.Upload) {
-	assert.Assert(t, mockProvider != nil)
-
-	mockScheme := "file://"
-
-	for range expectedUploads {
-		mockProvider.On("GetScheme").Return(&mockScheme).Once()
-		mockProvider.On("Delete", mock.Anything, mock.Anything, mock.Anything).Return(nil).Once()
-	}
-}
-
-// ExpectUploadNillable sets up the mock object store to expect an upload and related operations
-func ExpectUploadNillable(t *testing.T, mockProvider *mock_shared.MockProvider, expectedUploads []*graphql.Upload) {
-	assert.Check(t, mockProvider != nil)
-
-	mockScheme := "file://"
-
-	for _, upload := range expectedUploads {
-		if upload != nil {
-			mockProvider.On("GetScheme").Return(&mockScheme).Once()
-			mockProvider.On("ProviderType").Return(storage.DiskProvider).Maybe()
-			mockProvider.On("Upload", mock.Anything, mock.Anything, mock.Anything).Return(&storage.UploadedMetadata{
-				FileMetadata: pkgobjects.FileMetadata{
-					Key:          "test-key",
-					Size:         upload.Size,
-					Folder:       "test-folder",
-					Bucket:       "test-bucket",
-					ContentType:  upload.ContentType,
-					ProviderType: storage.DiskProvider,
-					FullURI:      "file:///tmp/test-file",
-				},
-			}, nil).Once()
-
-			// Allow document hooks to download the just-uploaded content for parsing
-			mockProvider.On("Download", mock.Anything, mock.Anything, mock.Anything).Return(&storage.DownloadedMetadata{
-				File: []byte("test content"),
-				Size: upload.Size,
-			}, nil).Maybe()
-		}
-	}
-}
-
-// ExpectUploadCheckOnly sets up the mock object store to expect an upload check only operation
-// but fails before the upload is attempted
-func ExpectUploadCheckOnly(t *testing.T, mockProvider *mock_shared.MockProvider) {
-	assert.Assert(t, mockProvider != nil)
-
-	mockScheme := "file://"
-
-	mockProvider.On("GetScheme").Return(&mockScheme).Once()
-}
-
-// ParseClientError parses the error response from the client and returns a slice of gqlerror.Error
-func ParseClientError(t *testing.T, err error) []*gqlerror.Error {
-	t.Helper()
-
-	if err == nil {
-		return nil
-	}
-
-	errResp, ok := err.(*clientv2.ErrorResponse)
-	assert.Check(t, ok)
-	assert.Check(t, errResp.HasErrors())
-
-	gqlErrors := []*gqlerror.Error{}
-
-	errors := errResp.GqlErrors.Unwrap()
-
-	for _, e := range errors {
-		customErr, ok := e.(*gqlerror.Error)
-		assert.Check(t, ok)
-		gqlErrors = append(gqlErrors, customErr)
-	}
-
-	return gqlErrors
-}
-
-// AssertErrorCode checks if the error code matches the expected code
-func AssertErrorCode(t *testing.T, err *gqlerror.Error, code string) {
-	t.Helper()
-
-	assert.Equal(t, code, testclient.GetErrorCode(err))
-}
-
-// AssertErrorMessage checks if the error message matches the expected message
-func AssertErrorMessage(t *testing.T, err *gqlerror.Error, msg string) {
-	t.Helper()
-
-	assert.Equal(t, msg, testclient.GetErrorMessage(err))
-}
-
-func RequireNoError(t *testing.T, err error) {
-	t.Helper()
-	if err != nil {
-		log.Error().Err(err).Msg("fatal error during test setup or teardown")
-
-		os.Exit(1)
-	}
-}
-
-func FailNow(t *testing.T, msgs ...string) {
-	t.Helper()
-	logMsg := log.Error()
-
-	for _, m := range msgs {
-		logMsg.Str("msg", m)
-	}
-
-	logMsg.Msg("fatal error during test setup or teardown")
-
-	os.Exit(1)
-}
-
-func TestPDFBytes() []byte {
-	page := map[string]any{
-		"paper":  "A4P",
-		"origin": "UpperLeft",
-		"fonts": map[string]any{
-			"f": map[string]any{"name": "Helvetica", "size": 12},
-		},
-		"pages": map[string]any{
-			"1": map[string]any{
-				"content": map[string]any{
-					"text": []map[string]any{
-						{"value": "test", "pos": [2]float64{20, 20}, "font": map[string]any{"name": "$f"}},
-					},
-				},
-			},
-		},
-	}
-
-	jsonData, _ := json.Marshal(page)
-
-	var buf bytes.Buffer
-	_ = api.Create(nil, bytes.NewReader(jsonData), &buf, nil)
-
-	return buf.Bytes()
-}
-
-// MockStripeClient creates a new stripe client with mock backend
-func (suite *GraphTestSuite) MockStripeClient() (*entitlements.StripeClient, error) {
-	suite.StripeMockBackend = new(mocks.MockStripeBackend)
-	stripeTestBackends := &stripe.Backends{
-		API:     suite.StripeMockBackend,
-		Connect: suite.StripeMockBackend,
-		Uploads: suite.StripeMockBackend,
-	}
-
-	suite.OrgSubscriptionMocks()
-
-	return entitlements.NewStripeClient(entitlements.WithAPIKey("sk_test_testing"),
-		entitlements.WithConfig(entitlements.Config{
-			Enabled:             true,
-			StripeWebhookSecret: webhookSecret,
-		},
-		),
-		entitlements.WithBackends(stripeTestBackends),
-	)
-}
-
-var MockItems = []*stripe.SubscriptionItem{
-	{
-		Price: &stripe.Price{
-			Product: &stripe.Product{
-				ID: "prod_test_product",
-			},
-			ID: "price_test_price",
-			Recurring: &stripe.PriceRecurring{
-				Interval: "month",
-			},
-			Currency:   "usd",
-			UnitAmount: 1000,
-		},
-	},
-}
-
-// MockCustomer for webhook tests
-var MockCustomer = &stripe.Customer{
-	ID: "cus_test_customer",
-	Subscriptions: &stripe.SubscriptionList{
-		Data: []*stripe.Subscription{
-			{
-				Customer: &stripe.Customer{
-					ID: "cus_test_customer",
-				},
-				ID: seedStripeSubscriptionID,
-				Items: &stripe.SubscriptionItemList{
-					Data: MockItems,
-				},
-			},
-		},
-	},
-}
-
-var MockSubscription = &stripe.Subscription{
-	ID:     "sub_test_subscription",
-	Status: "active",
-	Items: &stripe.SubscriptionItemList{
-		Data: MockItems,
-	},
-	Metadata: map[string]string{
-		"organization_id": ulids.New().String(),
-	},
-	Customer: &stripe.Customer{
-		ID: "cus_test_customer",
-	},
-	TrialEnd:     time.Now().Add(7 * 24 * time.Hour).Unix(), // 7 days from now
-	DaysUntilDue: 15,
-}
-
-var MockProduct = &stripe.Product{
-	ID:   "prod_test_product",
-	Name: "Test Product",
-}
-
-// OrgSubscriptionMocks mocks the stripe calls for org subscription during the webhook tests
-func (suite *GraphTestSuite) OrgSubscriptionMocks() {
-	// setup mocks for get customer by id
-	suite.StripeMockBackend.On("Call", mock.Anything, mock.Anything, mock.Anything, mock.AnythingOfType("*stripe.CustomerRetrieveParams"), mock.AnythingOfType("*stripe.Customer")).Run(func(args mock.Arguments) {
-		mockCustomerSearchResult := args.Get(4).(*stripe.Customer)
-
-		*mockCustomerSearchResult = *MockCustomer
-
-	}).Return(nil)
-
-	// setup mocks for creating customer params
-	suite.StripeMockBackend.On("Call", mock.Anything, mock.Anything, mock.Anything, mock.AnythingOfType("*stripe.CustomerCreateParams"), mock.AnythingOfType("*stripe.Customer")).Run(func(args mock.Arguments) {
-		mockCustomerSearchResult := args.Get(4).(*stripe.Customer)
-
-		*mockCustomerSearchResult = *MockCustomer
-
-	}).Return(nil)
-
-	// mock customer search
-	suite.StripeMockBackend.On("CallRaw", mock.Anything, mock.Anything, mock.Anything, mock.AnythingOfType("*stripe.Params"), mock.AnythingOfType("*stripe.v1SearchPage[*github.com/stripe/stripe-go/v86.Customer]")).Run(func(args mock.Arguments) {
-		out := args.Get(4) // this is *v1SearchPage[*stripe.Customer] now, but unexported
-
-		// Build a payload that matches Stripe search response shape
-		payload := map[string]any{
-			"object":   "search_result",
-			"data":     []*stripe.Customer{MockCustomer},
-			"has_more": false,
-		}
-
-		b, _ := json.Marshal(payload)
-		_ = json.Unmarshal(b, out)
-	}).Return(nil)
-
-	// mock for subscription create params
-	suite.StripeMockBackend.On("Call", mock.Anything, mock.Anything, mock.Anything, mock.AnythingOfType("*stripe.SubscriptionCreateParams"), mock.AnythingOfType("*stripe.Subscription")).Run(func(args mock.Arguments) {
-		mockSubscriptionSearchResult := args.Get(4).(*stripe.Subscription)
-
-		*mockSubscriptionSearchResult = *MockSubscription
-
-	}).Return(nil)
-
-	// mock for product retrieve params
-	suite.StripeMockBackend.On("Call", mock.Anything, mock.Anything, mock.Anything, mock.AnythingOfType("*stripe.ProductRetrieveParams"), mock.AnythingOfType("*stripe.Product")).Run(func(args mock.Arguments) {
-		mockProductRetrieveResult := args.Get(4).(*stripe.Product)
-
-		*mockProductRetrieveResult = *MockProduct
-
-	}).Return(nil)
-
-	// mock for product params
-	suite.StripeMockBackend.On("Call", mock.Anything, mock.Anything, mock.Anything, mock.AnythingOfType("*stripe.SubscriptionRetrieveParams"), mock.AnythingOfType("*stripe.Product")).Run(func(args mock.Arguments) {
-		mockSubscriptionRetrieveResult := args.Get(4).(*stripe.Subscription)
-
-		*mockSubscriptionRetrieveResult = *MockSubscription
-
-	}).Return(nil)
-
-	// setup mocks for getting entitlements
-	suite.StripeMockBackend.On("CallRaw", mock.Anything, mock.Anything, mock.Anything, mock.AnythingOfType("*stripe.Params"), mock.AnythingOfType("*stripe.EntitlementsActiveEntitlementList")).Run(func(args mock.Arguments) {
-		mockCustomerSearchResult := args.Get(4).(*stripe.EntitlementsActiveEntitlementList)
-
-		*mockCustomerSearchResult = stripe.EntitlementsActiveEntitlementList{
-			Data: []*stripe.EntitlementsActiveEntitlement{
-				{
-					Feature: &stripe.EntitlementsFeature{
-						ID:        "feat_test_feature",
-						LookupKey: "test_feature",
-					},
-				},
-			},
-		}
-
-	}).Return(nil)
-
-	// setup mocks for subscription schedule creation
-	suite.StripeMockBackend.On("Call", mock.Anything, mock.Anything, mock.Anything, mock.AnythingOfType("*stripe.SubscriptionScheduleCreateParams"), mock.AnythingOfType("*stripe.SubscriptionSchedule")).Run(func(args mock.Arguments) {
-		mockSubscriptionScheduleResult := args.Get(4).(*stripe.SubscriptionSchedule)
-
-		*mockSubscriptionScheduleResult = stripe.SubscriptionSchedule{
-			ID:     "sched_test_schedule",
-			Status: "active",
-		}
-
-	}).Return(nil)
-
-	// setup mocks for customer update params
-	suite.StripeMockBackend.On("Call", mock.Anything, mock.Anything, mock.Anything, mock.AnythingOfType("*stripe.CustomerUpdateParams"), mock.AnythingOfType("*stripe.Customer")).Run(func(args mock.Arguments) {
-		mockCustomerUpdateResult := args.Get(4).(*stripe.Customer)
-
-		*mockCustomerUpdateResult = *MockCustomer
-
-	}).Return(nil)
 }
 
 // NewTestGraphServer creates a new GraphQL server for testing
